inside/requirements.txt
asgiref==3.4.0
certifi==2021.5.30
Django==3.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
channels==3.0.3
channels-redis==3.2.0
requests==2.23.0
